+++ src/utils/Constants.js
@@ -1,0 +1,4 @@
+/* 
+    包含应用中一些常量名称的模块
+*/
+export const PAGE_SIZE = 2
